refac:Re-structure initial layout from App.js into /components dir and /screens dir under 'features/restaurants'
