+++ App.js
@@ -1,36 +1,13 @@
 import { StatusBar as ExpoStatusBar } from 'expo-status-bar';
 import React from 'react';
-import { StatusBar, StyleSheet, SafeAreaView, View, Text } from 'react-native';
-import { Searchbar } from 'react-native-paper';
+
+import { RestaurantsScreen } from './src/features/restaurants/screens/RestaurantsScreen';
 
 export default function App() {
   return (
     <>
-      <SafeAreaView style={styles.container}>
-        <View style={styles.searchBar}>
-          <Searchbar placeholder="Search" />
-        </View>
-        <View style={styles.list}>
-          <Text>Result list</Text>
-        </View>
-      </SafeAreaView>
+      <RestaurantsScreen />
       <ExpoStatusBar style="auto" />
     </>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    marginTop: StatusBar.currentHeight,
-  },
-  searchBar: {
-    padding: 20,
-    backgroundColor: '#33FF90',
-  },
-  list: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: '#33A8FF',
-  },
-});
